refactor: MVC 패턴과 유사하게 클래스 배치
