+++ src/com/example/kiosk/controller/Kiosk.java
@@ -1,4 +1,9 @@
-package com.example.kiosk;
+package com.example.kiosk.controller;
+
+import com.example.kiosk.service.Cart;
+import com.example.kiosk.dao.Menu;
+import com.example.kiosk.dto.Discount;
+import com.example.kiosk.dto.MenuItem;
 
 import java.io.BufferedReader;
 import java.io.IOException;
