+++ src/com/example/kiosk/dao/Menu.java
@@ -1,4 +1,6 @@
-package com.example.kiosk;
+package com.example.kiosk.dao;
+
+import com.example.kiosk.dto.MenuItem;
 
 import java.util.ArrayList;
 import java.util.List;
